+++ src/WeatherService/Service.hs
@@ -1,4 +1,5 @@
-{-# LANGUAGE OverloadedStrings, FlexibleContexts, TypeFamilies, QuasiQuotes, TemplateHaskell, DeriveGeneric #-}
+{-# LANGUAGE OverloadedStrings, FlexibleContexts, TypeFamilies
+  , QuasiQuotes, TemplateHaskell, DeriveGeneric #-}
 module WeatherService.Service (WeatherField(..)
                               , dayHandler
                               , dayPutHandler
@@ -44,10 +45,10 @@ dayHandler d conn = do
   r <- liftIO (queryNamed conn "SELECT the_date, temperature \
                                \ FROM  weather \
                                \ WHERE the_date = :dt" [":dt" := d] :: IO [WeatherField])
-  liftIO $ debugM "Date Query" (listToOutput r) -- ^ NB example of how to output debug messages
+  liftIO $ debugM "Date Query" $ listToOutput r -- ^ NB example of how to output debug messages
   case r of
     [] -> notFoundHandler
-    _  -> ok $ toResponse (listToOutput r)
+    _  -> ok $ toResponse $ listToOutput r
 
 {-| Handle PUT requests for date/temperature pairs. -}
 dayPutHandler :: Text -> Text -> Connection -> ServerPart Response
@@ -55,7 +56,7 @@ dayPutHandler d t conn = do
   r <- liftIO (queryNamed conn "SELECT the_date, temperature \
                                \ FROM  weather \
                                \ WHERE the_date = :dt" [":dt" := d] :: IO [WeatherField])
-  liftIO $ debugM "Date PUT request" (listToOutput r)
+  liftIO $ debugM "Date PUT request" $ listToOutput r
   case r of
     [] -> insertHandler d t conn
     _  -> updateHandler d t conn
@@ -65,21 +66,21 @@ insertHandler :: Text -> Text -> Connection -> ServerPart Response
 insertHandler d t conn = do
   let t' = (read $ unpack t)::Float
   liftIO (execute conn "INSERT INTO weather (the_date, temperature) VALUES (?,?)" (WeatherField d t'))
-  ok $ emptyJSONResponse
+  ok emptyJSONResponse
 
 {-| Update a date/temperature pair. -}
 updateHandler :: Text -> Text -> Connection -> ServerPart Response
 updateHandler d t conn = do
   liftIO (execute conn "UPDATE weather SET temperature = ? WHERE the_date = ?" (Only t))
-  ok $ emptyJSONResponse
+  ok emptyJSONResponse
 
 {-| Return 404 Not Found and an empty JSON object -}
 notFoundHandler :: ServerPart Response
-notFoundHandler = notFound $ emptyJSONResponse
+notFoundHandler = notFound emptyJSONResponse
 
 {-| An empty JSON object -}
 --emptyJSONObject :: ServerPart Response
-emptyJSONResponse = toResponse (pack "[]")
+emptyJSONResponse = toResponse $ pack "[]"
 
 {-| Turn a list of WeatherFields into a JSON object. -}
 listToOutput :: ToJSON a => [a] -> String
@@ -92,10 +93,10 @@ rangeHandler d1 d2 conn = do
                                \ FROM  weather \
                                \ WHERE the_date BETWEEN :dt1 AND :dt2"
                                [":dt1" := d1,":dt2" := d2] :: IO [WeatherField])
-  liftIO $ debugM "Range Query" (listToOutput r)
+  liftIO $ debugM "Range Query" $ listToOutput r
   case r of
     [] -> notFoundHandler
-    _  -> ok $ toResponse (listToOutput r)
+    _  -> ok $ toResponse $ listToOutput r
 
 {-| Handle requests for dates in between date 1 and date 2 with max temperature.-}
 maxHandler :: Text -> Text -> Connection -> ServerPart Response
@@ -104,18 +105,18 @@ maxHandler d1 d2 conn = do
                                \ FROM  weather \
                                \ WHERE the_date BETWEEN :dt1 AND :dt2"
                                [":dt1" := d1,":dt2" := d2] :: IO [WeatherField])
-  liftIO $ debugM "Max Query" (listToOutput r)
+  liftIO $ debugM "Max Query" $ listToOutput r
   case r of
     [] -> notFoundHandler
-    _  -> ok $ toResponse (listToOutput r)
+    _  -> ok $ toResponse $ listToOutput r
 
 {-| Handle requests for dates that has a higher temperature than t. -}
-aboveTHandler :: Float -> Connection -> ServerPart Response
+aboveTHandler :: Text -> Connection -> ServerPart Response
 aboveTHandler t conn = do
   r <- liftIO (queryNamed conn "SELECT the_date, temperature \
                                \ FROM  weather \
                                \ WHERE temperature >= :dt" [":dt" := t] :: IO [WeatherField])
-  liftIO $ debugM "Above Query" (listToOutput r)
+  liftIO $ debugM "Above Query" $ listToOutput r
   case r of
     [] -> notFoundHandler
-    _  -> ok $ toResponse (listToOutput r)
+    _  -> ok $ toResponse $ listToOutput r