+++ ChangeLog.md
@@ -3,3 +3,10 @@
 ## 0.1.0.0  -- YYYY-mm-dd
 
 * First version. Released on an unsuspecting world.
+
+## 0.1.5.0  -- 2018-04-24
+
+* Implement range/d1/d2
+* Implement max/d1/d2
+* Implement above/temperature
+* Response 405 "Method not allowed" on illigel request